Reset timeoutlen in vimrc
set timeoutlen=200 (previously 1000) because the was an annoying delay
when using <leader> in insert mode. It is strange that there was a delay
at all; I am not sure why vim is waiting. Reseting timeoutlen is a bad
solution, since it doesn't address the root problem and since it makes
the use of <leader> difficult.