GoYo gets its own vimrc
I use GoYo mode when I am writing English. When I do so, I don't need
any highlighting, buffer bars, or anything at all. Only GoYo and Pencil.
But turning all the other stuff off is a real pill. So I decided it
would be easier to just make a second vimrc and explicitly call it
(though an alias, of course, as defined in .bash_aliases).

A side effect of this change is that opening windows in nearly vanilla
vi is super fast.