Remove vi-mode and fix readline
The battery checking function I added earlier caused readline to
overwrite the current line, rather than wrap. This was because I was not
informing it about invisible characters. Invisible characters (like
colors) need to be bound by \x01 and \x02, which indicate start and stop
of invisible sequece.

Also, I am ditching vi-mode. Vi is imersive, not something to flicker
in and out of. And a half implemented vi is far worse than nothing.