Found a more permanent solution to my problem
For years I've been rebuilding vim from source every time it updated. I
did this because the the clipboard compiler flag was missing in the
default Arch Linux build. But I've finally found that if you install
gvim rather than vim, the clipboard flag is set correctly.

This still isn't a real solution. What I really need is a scripted
solution with full control over the compilation flags. But, it is at
least an improvement ...